Minor updates to libgeometry

+++ libgeometry/src/point.c
@@ -70,14 +70,10 @@ int point_orientation(point_t *p, point_t *q, point_t *r)
 {
     // See https://www.geeksforgeeks.org/orientation-3-ordered-points/
     // for details of below formula.
-    int val = (q->y - p->y) * (r->x - q->x) -
-              (q->x - p->x) * (r->y - q->y);
+    double val = (q->y - p->y) * (r->x - q->x) -
+                 (q->x - p->x) * (r->y - q->y);
 
-    if (val == 0) return 0;  // colinear
+    if (val == 0.0) return 0;  // colinear
 
-    return (val > 0)? 1: 2; // clock or counterclock wise
+    return (val > 0.0)? 1: 2; // clock or counterclock wise
 }
-
-
-
-
